+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2026-07-21]
+## [1.1.0 - 2026-07-22]
 ### Added
 - Dockerfile
 - Single Spa
@@ -152,6 +152,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Cohort definition in a query
 - Query Details
 - Rename overview headers
+- Filter services and auto select service if only one
 
 ### Removed
 - Application Form